# perf(namespace): speed up Namespace comparisons